requirements/common.txt
Django==4.1.5
django-cors-headers==3.13.0
python-decouple==3.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
requests==2.28.2
tzdata